Refactor: Add AddTransactionBloc for managing transaction creation state
This commit introduces a new BLoC, `AddTransactionBloc`, to handle the state management for the "add transaction" screen. This refactoring separates the concerns of transaction creation logic from the UI, leading to a cleaner and more maintainable codebase.

The `AddTransactionBloc` is responsible for:
- Loading available categories based on the selected money type.
- Managing the selected date for the transaction.
- Handling the selection of a category.
- Providing the necessary state (selected date, money type, selected category, available categories) to the UI.

The `AddTransactionScreen` has been updated to use this new BLoC, delegating state management tasks to it. This includes:
- Initializing the BLoC with the initial money type and category.
- Dispatching events to the BLoC for loading categories, selecting a date, and selecting a category.
- Building the UI based on the state emitted by the BLoC.

Additionally, the transaction list screen (`TransactionScreen`) header was simplified by removing the notification icon, as it was deemed redundant with the notification icon present in the `AddTransactionScreen`'s app bar. The `TransactionBloc` and related files have been updated to reflect changes in model imports due to the refactoring.
